+++ EveOnlineIskPerHour/WalletLog/WalletLogCalculationResult.cs
@@ -2,7 +2,7 @@
 
 namespace EveOnlineIskPerHour.WalletLog
 {
-    public record WalletLogCalculationResult(int IskPerHour, int TotalIncome, TimeSpan TotalTime, 
+    public record WalletLogCalculationResult(double IskPerHour, double TotalIncome, TimeSpan TotalTime, 
         TimeSpan BestSiteTime, TimeSpan AverageSiteTime, TimeSpan WorstSiteTime, 
         int MeasuredSites);
